Update reservation, schedule and occurrence models, repositories and utils

+++ server/repositories/occurrence_repository.py
@@ -21,8 +21,9 @@ class OccurrenceRepository:
         occurrences = OccurrenceUtils.occurrences_from_schedules(schedule)
 
         previous_occurrences = schedule.occurrences
-        for occurrence in previous_occurrences:
-            session.delete(occurrence)
+        if previous_occurrences:
+            for occurrence in previous_occurrences:
+                session.delete(occurrence)
 
         schedule.occurrences = occurrences
         classroom.occurrences.extend(occurrences)
@@ -39,8 +40,9 @@ class OccurrenceRepository:
 
     @staticmethod
     def remove_schedule_allocation(schedule: Schedule, session: Session) -> None:
-        for occurrence in schedule.occurrences:
-            session.delete(occurrence)
+        if schedule.occurrences:
+            for occurrence in schedule.occurrences:
+                session.delete(occurrence)
         schedule.allocated = False
         session.add(schedule)
